Add highscore tracking functionality

+++ number_ninja/number_ninja.rb
@@ -3,11 +3,12 @@ class NumberNinja
     @level = 0
     @guesses = 0
     @points = 0
+    @timetrack = false
+    @t = nil
   end
 
   # Display a welcome message
   # @return [void]
-  # DONE
   def hello
     puts "Willkommen bei NumberNinja!"
 
@@ -28,13 +29,13 @@ class NumberNinja
 
   # Load the highscore from a file
   # @return [void]
-  # DONE
   def loadHighscore
     File.open("data/highscore.txt", "r") do |file|
       file.each_line do |line|
         file_data = line.split(";")
         @level = file_data[1].to_i
-        @points = file_data[2].to_i
+        @timetrack = file_data[2] == "true"
+        @points = file_data[3].to_i
       end
     end
     File.delete("data/highscore.txt")
@@ -44,11 +45,14 @@ class NumberNinja
 
   # Start a new game
   # @return [void]
-  # DONE
   def newGame
     puts "Neues Spiel starten!"
     print "Wähle eine Schwierigkeitsstufe: (1) Leicht 0-50, (2) Mittel 0-100, (3) Schwer 0-200: \n"
     difficulty = gets.chomp.to_i
+    print "Zeitlimit (30s) aktivieren? (j/n): "
+    timetrack = gets.chomp.to_s
+    @timetrack = (timetrack == "j")
+
     case difficulty
     when 1
       @level = 1
@@ -66,7 +70,6 @@ class NumberNinja
 
   # Play the game
   # @return [void]
-  # DONE
   def play
     case @level
     when 1
@@ -78,16 +81,26 @@ class NumberNinja
     end
 
     puts "Errate die Zahl zwischen 0 und " + maxnum.to_s + "!"
-
     @secret_number = rand(0..maxnum)
+
+    if @timetrack
+      @start_time = Time.now
+      puts "Du hast 30 Sekunden Zeit!"
+      @t = Thread.new do
+        sleep 30
+        puts "\n\nZeit abgelaufen!"
+        play_again
+      end
+    end
 
     loop do
       print "Gib deine Vermutung ein: "
       guess = gets.chomp.to_i
-      @guesses = @guesses + 1
+      @guesses += 1
 
       if validate(guess, maxnum)
         if guess == @secret_number
+          @t.kill if @timetrack && @t
           win
           break
         elsif guess < @secret_number
@@ -99,47 +112,51 @@ class NumberNinja
     end
   end
 
-  # Start the game
-  # @return [void]
-  # DONE
+  # Validate the guess
+  # @return [Boolean]
   def validate(guess, maxnum)
     if guess >= 0 && guess <= maxnum
-      return true
+      true
     else
-      puts "Felerhafte Eingabe: Zahl nicht zwischen 0 und " + maxnum.to_s + "."
-      return false
+      puts "Fehlerhafte Eingabe: Zahl nicht zwischen 0 und " + maxnum.to_s + "."
+      false
     end
   end
 
   # Display a win message
   # @return [void]
-  # DONE
   def win
     puts "Glückwunsch! Du hast die Zahl erraten!"
     puts "Du hast " + @guesses.to_s + " Versuche gebraucht!"
 
     if @guesses == 1
-      @points = @points + 1000
+      @points += 1000
       puts "Perfekt! Du erhältst 1000 Punkte!"
     elsif @guesses < 5
-      @points = @points + 500
+      @points += 500
       puts "Schnell geraten! Du erhältst 500 Punkte!"
     elsif @guesses < 10
-      @points = @points + 200
+      @points += 200
       puts "Gut gemacht! Du erhältst 200 Punkte!"
     elsif @guesses < 15
-      @points = @points + 100
+      @points += 100
       puts "Nicht schlecht! Du erhältst 100 Punkte!"
     else
       puts "Du erhältst keine Punkte."
     end
-    print "\n Deine Punktzahl: " + @points.to_s + "\n"
+    puts "\nDeine Punktzahl: " + @points.to_s + "\n"
 
+    play_again
+  end
+
+  # Prompt for replay
+  # @return [void]
+  def play_again
     print "Nochmal spielen? (j/n): "
     play_again = gets.chomp
     if play_again == "j"
       @guesses = 0
-      start
+      play
     else
       endRound
     end
@@ -147,7 +164,6 @@ class NumberNinja
 
   # End the round
   # @return [void]
-  # DONE
   def endRound
     print "Möchtest du deinen Highscore speichern? (j/n): "
     save = gets.chomp
@@ -160,22 +176,17 @@ class NumberNinja
 
   # Save the highscore to a file
   # @return [void]
-  # DONE
   def saveHighscore
-    if !File.exist?("data")
-      Dir.mkdir("data")
-    end
-    if File.exist?("data/highscore.txt")
-      File.delete("data/highscore.txt")
-    end
+    Dir.mkdir("data") unless File.exist?("data")
+    File.delete("data/highscore.txt") if File.exist?("data/highscore.txt")
+
     File.open("data/highscore.txt", "w") do |file|
-      file.write(Time.now.strftime("%Y-%m-%d %H:%M:%S"))
-      file.write(";")
-      file.write(@level)
-      file.write(";")
-      file.write(@points)
+      file.write(Time.now.strftime("%Y-%m-%d %H:%M:%S") + ";")
+      file.write(@level.to_s + ";")
+      file.write(@timetrack.to_s + ";")
+      file.write(@points.to_s)
     end
-    puts "Auf Wiedersehen!"
+    puts "Highscore gespeichert! Auf Wiedersehen!"
   end
 end
